requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
et-xmlfile==1.1.0
idna==3.2
importlib-metadata==4.6.1
joblib==1.0.1
nltk==3.6.2
numpy==1.19.5
openpyxl==3.0.7
pandas==1.1.5
Pillow==8.4.0
python-dateutil==2.8.2
pytz==2021.1
regex==2021.7.6
requests==2.26.0
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
threadpoolctl==2.2.0
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
xlrd==2.0.1
XlsxWriter==1.4.4
zipp==3.5.0
